setup_infra: add the /dev/shm entry with "noauto"
Tracked with Andriy - /dev/shm is not mounted via the normal
"mount -a" for fstab, but handled at different places in the
beginning of the boot. So in some cases running "mount -a" creates
a double mount, which leads to problems. Adding "noauto" removes
the "mount -a" issue.

Note that this doesn't fully prevents double mounts, but
we don't seem to have a way to do that without kernel changes.